Refactor: Modularize backend code into separate concerns
- Split app.py (506 lines) into focused modules:
  - config_manager.py: Configuration management with file persistence
  - api.py: API endpoints for chat proxy and configuration management
  - server.py: Flask app factory and frontend routing
  - app.py: Minimal entry point (8 lines)

- Improved code organization and maintainability
- Better separation of concerns
- Easier testing and future development
- All functionality preserved and tested working
